+++ include/qunit.hpp
@@ -31,7 +31,7 @@ class QUnit : public QParity, public QInterface {
 #endif
 protected:
     std::vector<QInterfaceEngine> engines;
-    int devID;
+    int64_t devID;
     QEngineShardMap shards;
     complex phaseFactor;
     bool doNormalize;
@@ -41,21 +41,21 @@ protected:
     bool isReactiveSeparate;
     bitLenInt thresholdQubits;
     real1_f separabilityThreshold;
-    std::vector<int> deviceIDs;
+    std::vector<int64_t> deviceIDs;
 
     QInterfacePtr MakeEngine(bitLenInt length, bitCapInt perm);
 
 public:
     QUnit(std::vector<QInterfaceEngine> eng, bitLenInt qBitCount, bitCapInt initState = 0U,
         qrack_rand_gen_ptr rgp = nullptr, complex phaseFac = CMPLX_DEFAULT_ARG, bool doNorm = false,
-        bool randomGlobalPhase = true, bool useHostMem = false, int deviceId = -1, bool useHardwareRNG = true,
-        bool useSparseStateVec = false, real1_f norm_thresh = REAL1_EPSILON, std::vector<int> devIDs = {},
+        bool randomGlobalPhase = true, bool useHostMem = false, int64_t deviceId = -1, bool useHardwareRNG = true,
+        bool useSparseStateVec = false, real1_f norm_thresh = REAL1_EPSILON, std::vector<int64_t> devIDs = {},
         bitLenInt qubitThreshold = 0U, real1_f separation_thresh = FP_NORM_EPSILON_F);
 
     QUnit(bitLenInt qBitCount, bitCapInt initState = 0U, qrack_rand_gen_ptr rgp = nullptr,
         complex phaseFac = CMPLX_DEFAULT_ARG, bool doNorm = false, bool randomGlobalPhase = true,
-        bool useHostMem = false, int deviceId = -1, bool useHardwareRNG = true, bool useSparseStateVec = false,
-        real1_f norm_thresh = REAL1_EPSILON, std::vector<int> devIDs = {}, bitLenInt qubitThreshold = 0U,
+        bool useHostMem = false, int64_t deviceId = -1, bool useHardwareRNG = true, bool useSparseStateVec = false,
+        real1_f norm_thresh = REAL1_EPSILON, std::vector<int64_t> devIDs = {}, bitLenInt qubitThreshold = 0U,
         real1_f separation_thresh = FP_NORM_EPSILON_F)
         : QUnit({ QINTERFACE_STABILIZER_HYBRID }, qBitCount, initState, rgp, phaseFac, doNorm, randomGlobalPhase,
               useHostMem, deviceId, useHardwareRNG, useSparseStateVec, norm_thresh, devIDs, qubitThreshold,
@@ -69,17 +69,17 @@ public:
     {
         QInterface::SetConcurrency(threadsPerEngine);
         ParallelUnitApply(
-            [](QInterfacePtr unit, real1_f unused1, real1_f unused2, real1_f unused3, int32_t threadsPerEngine) {
+            [](QInterfacePtr unit, real1_f unused1, real1_f unused2, real1_f unused3, int64_t threadsPerEngine) {
                 unit->SetConcurrency(threadsPerEngine);
                 return true;
             },
-            ZERO_R1_F, ZERO_R1_F, ZERO_R1_F, threadsPerEngine);
+            ZERO_R1_F, ZERO_R1_F, ZERO_R1_F, (uint32_t)threadsPerEngine);
     }
 
     virtual void SetReactiveSeparate(bool isAggSep) { isReactiveSeparate = isAggSep; }
     virtual bool GetReactiveSeparate() { return isReactiveSeparate; }
 
-    virtual void SetDevice(int dID);
+    virtual void SetDevice(int64_t dID);
     virtual int64_t GetDevice() { return devID; }
 
     virtual void SetQuantumState(const complex* inputState);
@@ -358,9 +358,9 @@ protected:
     virtual QInterfacePtr EntangleInCurrentBasis(
         std::vector<bitLenInt*>::iterator first, std::vector<bitLenInt*>::iterator last);
 
-    typedef bool (*ParallelUnitFn)(QInterfacePtr unit, real1_f param1, real1_f param2, real1_f param3, int32_t param4);
+    typedef bool (*ParallelUnitFn)(QInterfacePtr unit, real1_f param1, real1_f param2, real1_f param3, int64_t param4);
     bool ParallelUnitApply(ParallelUnitFn fn, real1_f param1 = ZERO_R1_F, real1_f param2 = ZERO_R1_F,
-        real1_f param3 = ZERO_R1_F, int32_t param4 = 0);
+        real1_f param3 = ZERO_R1_F, int64_t param4 = 0);
 
     virtual bool SeparateBit(bool value, bitLenInt qubit);
